Linking Parser to GUI and created the MainWindow.
Parser sources are now correctly compiled, linked in
the build system. Currently the GUI is built for WIN32,
but it's easy to port into Linux application.
Gui is linked against the Qt5 Widgets, Core respectively.

Gui is built against 32bit Qt5.8 because it's much
easier to work with those restraints than 64bit.
It would require more work todo than the deadline is.

Renamed WebStream to DataStream as it seems more to
provide data for the parser to stream than streaming
the Web.